services/core-api/requirements.txt
# Core API Dependencies

# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
tortoise-orm[asyncpg]==0.20.0
asyncpg==0.29.0
aerich==0.7.2

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
phonenumbers==8.13.26

# Date & Time
python-dateutil==2.8.2

# File Processing
openpyxl==3.1.2
pandas==2.1.4

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Scheduling
apscheduler==3.10.4

# Testing (development)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1